requirements.txt
openenv-core == 0.1.1
fastapi
uvicorn
pydantic
openai
google-generativeai